refactor: moved utils to another file
