+++ android/src/main/java/expo/modules/dragdropcontentview/ExpoDragDropContentView.kt
@@ -1,5 +1,6 @@
 package expo.modules.dragdropcontentview
 
+import android.annotation.SuppressLint
 import android.content.ClipData
 import android.content.ClipDescription.MIMETYPE_TEXT_PLAIN
 import android.content.ContentResolver
@@ -15,6 +16,7 @@ import android.util.Base64
 import android.util.Log
 import android.view.View
 import android.widget.Toast
+import androidx.annotation.RequiresApi
 import androidx.core.content.ContextCompat
 import androidx.core.view.DragStartHelper
 import androidx.draganddrop.DropHelper
@@ -24,12 +26,15 @@ import expo.modules.kotlin.views.ExpoView
 import java.io.ByteArrayOutputStream
 import java.io.File
 
+@SuppressLint("ViewConstructor")
 class ExpoDragDropContentView(context: Context, appContext: AppContext) : ExpoView(context, appContext) {
     private var includeBase64 = false
     private var draggableImageUris: List<String> = emptyList()
     private var highlightColor = ContextCompat.getColor(context, R.color.highlight_color)
     private var highlightBorderRadius = 0
     private val onDropEvent by EventDispatcher()
+
+    private val utils = Utils()
 
     fun setIncludeBase64(value: Boolean?) {
         includeBase64 = value ?: false
@@ -53,106 +58,9 @@ class ExpoDragDropContentView(context: Context, appContext: AppContext) : ExpoVi
         }
     }
 
-    private fun getImageDimensions(contentResolver: ContentResolver, contentUri: Uri): Pair<Int, Int> {
-        val options = BitmapFactory.Options()
-        options.inJustDecodeBounds = true
-
-        try {
-            // Use BitmapFactory to obtain image dimensions without loading the entire image into memory
-            BitmapFactory.decodeStream(contentResolver.openInputStream(contentUri), null, options)
-            return Pair(options.outWidth, options.outHeight)
-        } catch (e: Exception) {
-            e.printStackTrace()
-        }
-
-        return Pair(0, 0)
-    }
-    private fun showToast(message: String) {
-        Toast.makeText(this.context, message, Toast.LENGTH_SHORT).show()
-    }
-    private fun getFileInfo(contentResolver: ContentResolver, contentUri: Uri): Map<String, Any?>? {
-        val projection = arrayOf(
-            MediaStore.Images.Media.MIME_TYPE,
-            MediaStore.Images.Media.DISPLAY_NAME,
-            MediaStore.Images.Media.DATA
-        )
-
-        val cursor = contentResolver.query(contentUri, projection, null, null, null)
-
-        if (cursor == null) {
-            showToast("Not supported")
-        }
-
-        cursor?.use { cursorInstance ->
-            if (cursorInstance.moveToFirst()) {
-                val type = cursorInstance.getString(cursorInstance.getColumnIndexOrThrow(MediaStore.Images.Media.MIME_TYPE))
-                val fileName = cursorInstance.getString(cursorInstance.getColumnIndexOrThrow(MediaStore.Images.Media.DISPLAY_NAME))
-                val uri = "file://" + cursorInstance.getString(cursorInstance.getColumnIndexOrThrow(MediaStore.Images.Media.DATA))
-                val dimensions = getImageDimensions(contentResolver, contentUri)
-                val base64 = if (includeBase64) getBase64Data(contentResolver, contentUri) else null
-                val path = contentUri.path?.replace("/-1/1/", "")
-
-                val fileInfoMap = mutableMapOf(
-                    "width" to dimensions.first,
-                    "height" to dimensions.second,
-                    "type" to type,
-                    "fileName" to fileName,
-                    "uri" to uri,
-                    "path" to path
-                )
-
-                // Conditionally add base64 to the map if it's not null
-                base64?.let { fileInfoMap["base64"] = it }
-
-                return fileInfoMap
-            }
-        }
-
-
-        return null
-    }
-
-    private fun getBase64Data(contentResolver: ContentResolver, contentUri: Uri): String? {
-        try {
-            // Use BitmapFactory to decode the image
-            val bitmap = BitmapFactory.decodeStream(contentResolver.openInputStream(contentUri))
-
-            // Use a ByteArrayOutputStream to compress the bitmap to a byte array
-            val byteArrayOutputStream = ByteArrayOutputStream()
-            bitmap.compress(Bitmap.CompressFormat.JPEG, 100, byteArrayOutputStream)
-            val byteArray = byteArrayOutputStream.toByteArray()
-
-            // Use Base64 to encode the byte array to a base64 string
-            return Base64.encodeToString(byteArray, Base64.DEFAULT)
-        } catch (e: Exception) {
-            e.printStackTrace()
-        }
-
-        return null
-    }
-
-    private fun getContentUriForFile(context: Context, file: File): Uri? {
-        val projection = arrayOf(MediaStore.Images.Media._ID)
-        val selection = "${MediaStore.Images.Media.DATA} = ?"
-        val selectionArgs = arrayOf(file.absolutePath)
-        val sortOrder = null // You can specify sorting order if needed
-
-        val queryUri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI
-
-        context.contentResolver.query(queryUri, projection, selection, selectionArgs, sortOrder)?.use { cursor ->
-            if (cursor.moveToFirst()) {
-                val columnIndex = cursor.getColumnIndexOrThrow(MediaStore.Images.Media._ID)
-                val imageId = cursor.getLong(columnIndex)
-                return ContentUris.withAppendedId(queryUri, imageId)
-            }
-        }
-
-        return null
-    }
-
     private fun configureDropHelper() {
         // DropHelper is only available on Android N and above
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N) return
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N) return;
 
         val activity = appContext.activityProvider?.currentActivity!!
         val contentResolver = context.contentResolver
@@ -172,13 +80,20 @@ class ExpoDragDropContentView(context: Context, appContext: AppContext) : ExpoVi
             for (i in 0 until clipData.itemCount) {
                 val contentUri = clipData.getItemAt(i).uri
                 if (contentUri != null) {
-                    val info = getFileInfo(contentResolver, contentUri)
+                    val info = utils.getFileInfo(contentResolver, contentUri, includeBase64, this.context)
                     info?.let { infoList.add(it) }
                 }
             }
             if (infoList.isNotEmpty()) onDropEvent(mapOf("assets" to infoList))
             return@configureView null
         }
+    }
+
+    private fun configureDragHelper() {
+        // DropHelper is only available on Android N and above
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.N) return;
+
+        val contentResolver = context.contentResolver
 
         DragStartHelper(this) { view, _ ->
             val data: MutableList<Uri> = mutableListOf()
@@ -188,12 +103,11 @@ class ExpoDragDropContentView(context: Context, appContext: AppContext) : ExpoVi
 
                 if (!path.isNullOrBlank()) {
                     val file = File(path)
-                    val uri = getContentUriForFile(this.context, file)
+                    val uri = utils.getContentUriForFile(this.context, file)
                     uri?.let { data.add(it) }
                 }
             }
 
-            // Create a drag shadow builder
             val shadow = DragShadowBuilder(view)
 
             if (data.isNotEmpty()) {
@@ -202,17 +116,17 @@ class ExpoDragDropContentView(context: Context, appContext: AppContext) : ExpoVi
                 for (i in 1 until data.size) {
                     clipData.addItem(ClipData.Item(data[i]))
                 }
-                // Start the drag and drop operation
+
                 view.startDragAndDrop(clipData, shadow, null, DRAG_FLAG_GLOBAL or DRAG_FLAG_GLOBAL_URI_READ)
             }
-           else {
-                val clipData2 = ClipData.newUri(contentResolver, "Image", Uri.parse(""))
-                view.startDragAndDrop(clipData2, shadow, null, DRAG_FLAG_GLOBAL or DRAG_FLAG_GLOBAL_URI_READ)
+            else {
+                return@DragStartHelper false
             }
         }.attach()
     }
 
     init {
         configureDropHelper()
+        configureDragHelper()
     }
 }
